feature flag - registry deployments config file
flyway - delete all data at startup config property

+++ core/src/main/java/org/bf2/srs/fleetmanager/rest/service/impl/RegistryDeploymentServiceImpl.java
@@ -6,17 +6,33 @@ import org.bf2.srs.fleetmanager.rest.service.RegistryDeploymentService;
 import org.bf2.srs.fleetmanager.rest.service.convert.ConvertRegistryDeployment;
 import org.bf2.srs.fleetmanager.rest.service.model.RegistryDeployment;
 import org.bf2.srs.fleetmanager.rest.service.model.RegistryDeploymentCreate;
+import org.bf2.srs.fleetmanager.rest.service.model.RegistryDeploymentsConfigList;
 import org.bf2.srs.fleetmanager.storage.RegistryDeploymentNotFoundException;
 import org.bf2.srs.fleetmanager.storage.ResourceStorage;
 import org.bf2.srs.fleetmanager.storage.StorageConflictException;
 import org.bf2.srs.fleetmanager.storage.sqlPanacheImpl.model.RegistryDeploymentData;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
+
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+import javax.validation.ConstraintViolation;
+import javax.validation.ConstraintViolationException;
 import javax.validation.Valid;
+import javax.validation.Validator;
+import javax.ws.rs.ForbiddenException;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.toList;
 
@@ -29,6 +45,9 @@ public class RegistryDeploymentServiceImpl implements RegistryDeploymentService 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
     @Inject
+    Validator validator;
+
+    @Inject
     TaskManager tasks;
 
     @Inject
@@ -37,14 +56,81 @@ public class RegistryDeploymentServiceImpl implements RegistryDeploymentService 
     @Inject
     ConvertRegistryDeployment convertRegistryDeployment;
 
+    @ConfigProperty(name = "registry.deployments.config.file")
+    Optional<File> deploymentsConfigFile;
+
+    @Override
+    public void init() throws StorageConflictException, IOException {
+
+        if (deploymentsConfigFile.isEmpty()) {
+            return;
+        }
+
+        log.info("Loading registry deployments config file from {}", deploymentsConfigFile.get().getAbsolutePath());
+
+        YAMLMapper mapper = new YAMLMapper();
+
+        RegistryDeploymentsConfigList deploymentsConfigList = mapper.readValue(deploymentsConfigFile.get(), RegistryDeploymentsConfigList.class);
+
+        List<RegistryDeploymentCreate> staticDeployments = deploymentsConfigList.getDeployments();
+
+        Set<String> names = new HashSet<>();
+        List<String> duplicatedNames = staticDeployments.stream()
+                .map(d-> {
+                    Set<ConstraintViolation<RegistryDeploymentCreate>> errors = validator.validate(d);
+                    if (!errors.isEmpty()) {
+                        throw new ConstraintViolationException(errors);
+                    }
+                    return d;
+                })
+                .filter(d -> !names.add(d.getName()))
+                .map(d -> d.getName())
+                .collect(Collectors.toList());
+        if (!duplicatedNames.isEmpty()) {
+            throw new IllegalArgumentException("Error in static deployments config, duplicated deployments name: " + duplicatedNames.toString());
+        }
+
+        Map<String, RegistryDeploymentData> currentDeployments = storage.getAllRegistryDeployments().stream()
+                .collect(Collectors.toMap(d -> d.getName(), d -> d));
+
+        for (RegistryDeploymentCreate dep : staticDeployments) {
+
+            RegistryDeploymentData deploymentData = currentDeployments.get(dep.getName());
+
+            if (deploymentData == null) {
+                //deployment is new
+                deploymentData = convertRegistryDeployment.convert(dep);
+            } else {
+                RegistryDeploymentData configured = convertRegistryDeployment.convert(dep);
+                if (deploymentData.getRegistryDeploymentUrl().equals(configured.getRegistryDeploymentUrl())
+                        && deploymentData.getTenantManagerUrl().equals(configured.getTenantManagerUrl())) {
+                    //no changes in the deployment
+                    continue;
+                }
+
+                deploymentData.setRegistryDeploymentUrl(dep.getRegistryDeploymentUrl());
+                deploymentData.setTenantManagerUrl(dep.getTenantManagerUrl());
+            }
+
+            createOrUpdateRegistryDeployment(deploymentData);
+        }
+    }
+
     @Override
     public RegistryDeployment createRegistryDeployment(@Valid RegistryDeploymentCreate deploymentCreate) throws StorageConflictException {
-        //TODO validate values
-        //registryDeploymentURl finishes without / starts with http ...
+        if (deploymentsConfigFile.isPresent()) {
+            throw new ForbiddenException();
+        }
         RegistryDeploymentData deployment = convertRegistryDeployment.convert(deploymentCreate);
-        storage.createOrUpdateRegistryDeployment(deployment);
-        tasks.submit(RegistryDeploymentHeartbeatTask.builder().deploymentId(deployment.getId()).build());
+        createOrUpdateRegistryDeployment(deployment);
         return convertRegistryDeployment.convert(deployment);
+    }
+
+    private void createOrUpdateRegistryDeployment(RegistryDeploymentData deployment) throws StorageConflictException {
+        boolean created = storage.createOrUpdateRegistryDeployment(deployment);
+        if (created) {
+            tasks.submit(RegistryDeploymentHeartbeatTask.builder().deploymentId(deployment.getId()).build());
+        }
     }
 
     @Override
@@ -63,6 +149,9 @@ public class RegistryDeploymentServiceImpl implements RegistryDeploymentService 
 
     @Override
     public void deleteRegistryDeployment(Long id) throws RegistryDeploymentNotFoundException, StorageConflictException {
+        if (deploymentsConfigFile.isPresent()) {
+            throw new ForbiddenException();
+        }
         storage.deleteRegistryDeployment(id);
     }
 }
